requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
phonenumbers==8.12.13
Pillow==2.2.1
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
